backend/requirements.txt
langgraph>=0.2.6
langchain>=0.3.19
langchain-core>=0.3.0
langchain-google-genai>=0.1.0
python-dotenv>=1.0.1
langgraph-sdk>=0.1.57
langgraph-cli>=0.1.71
langgraph-api>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
google-genai>=0.3.0
pydantic>=2.0.0
typing-extensions>=4.0.0
starlette>=0.27.0
pymongo>=4.6.0
